Restructure project for system-only integration branch
