+++ src/analytics_component/README.md
@@ -1,7 +1,8 @@
 ## Компонент аналитики
 
-Логический компонент для операции:
+Компонент обрабатывает операцию:
 
 - `get_analytics`
 
-Реализация пока остаётся in-process и вызывается через gateway.
+Код обработки входящих bus-сообщений находится в `component.go`, а вызов
+компонента выполняется через gateway.
