+++ src/contracts_component/README.md
@@ -1,10 +1,11 @@
 ## Компонент контрактов
 
-Логический компонент для операций:
+Компонент обрабатывает операции:
 
 - `conclude_contract`
 - `confirm_execution`
 - `create_dispute`
 
-Реализация пока остаётся in-process, а отдельная директория компонента нужна
-для совместимости интеграционной ветки с требуемой структурой системы.
+Код обработки входящих bus-сообщений находится в `component.go`.
+Компонент вызывается через gateway, а общий слой совместимости остаётся в
+`internal/handler`.
