+++ src/contracts_component/component.go
@@ -1,6 +1,13 @@
 package contracts_component
 
-import "github.com/kirilltahmazidi/aggregator/internal/models"
+import (
+	"encoding/json"
+	"log"
+	"time"
+
+	"github.com/google/uuid"
+	"github.com/kirilltahmazidi/aggregator/internal/models"
+)
 
 const Topic = "components.agregator.contracts"
 
@@ -17,4 +24,87 @@ func Handles(action models.MessageType) bool {
 		}
 	}
 	return false
+}
+
+type Handler struct{}
+
+func NewHandler() *Handler {
+	return &Handler{}
+}
+
+func (h *Handler) Handle(req models.Request) (models.Response, bool) {
+	switch req.Action {
+	case models.MsgConcludeContract:
+		return h.concludeContract(req), true
+	case models.MsgConfirmExecution:
+		return h.confirmExecution(req), true
+	case models.MsgCreateDispute:
+		return h.createDispute(req), true
+	default:
+		return models.Response{}, false
+	}
+}
+
+func (h *Handler) concludeContract(req models.Request) models.Response {
+	var payload models.ConcludeContractRequest
+	if err := json.Unmarshal(req.Payload, &payload); err != nil {
+		return errResponse(req, "invalid payload: "+err.Error())
+	}
+
+	return okResponse(req, models.ConcludeContractResponse{
+		ContractID: uuid.NewString(),
+		OrderID:    payload.OrderID,
+		Status:     "active",
+	})
+}
+
+func (h *Handler) confirmExecution(req models.Request) models.Response {
+	var payload models.ConfirmExecutionRequest
+	if err := json.Unmarshal(req.Payload, &payload); err != nil {
+		return errResponse(req, "invalid payload: "+err.Error())
+	}
+
+	return okResponse(req, models.ConfirmExecutionResponse{
+		ContractID: payload.ContractID,
+		Status:     "completed",
+		Message:    "contract marked as completed by customer (stub)",
+	})
+}
+
+func (h *Handler) createDispute(req models.Request) models.Response {
+	var payload models.CreateDisputeRequest
+	if err := json.Unmarshal(req.Payload, &payload); err != nil {
+		return errResponse(req, "invalid payload: "+err.Error())
+	}
+
+	return okResponse(req, models.CreateDisputeResponse{
+		DisputeID:       uuid.NewString(),
+		ContractID:      payload.ContractID,
+		Status:          "dispute_opened",
+		InsurancePayout: payload.ClaimAmount,
+		Message:         "dispute registered, insurance payout initiated (stub)",
+	})
+}
+
+func okResponse(req models.Request, payload interface{}) models.Response {
+	return models.Response{
+		Action:        models.ResponseAction,
+		Payload:       payload,
+		Sender:        models.DefaultSender,
+		CorrelationID: req.GetCorrelationID(),
+		Success:       true,
+		Timestamp:     time.Now().UTC().Format(time.RFC3339Nano),
+	}
+}
+
+func errResponse(req models.Request, msg string) models.Response {
+	log.Printf("[contracts_component] error correlation_id=%s: %s", req.GetCorrelationID(), msg)
+	return models.Response{
+		Action:        models.ResponseAction,
+		Sender:        models.DefaultSender,
+		CorrelationID: req.GetCorrelationID(),
+		Success:       false,
+		Error:         msg,
+		Timestamp:     time.Now().UTC().Format(time.RFC3339Nano),
+	}
 }
