+++ src/orders_component/README.md
@@ -1,11 +1,11 @@
 ## Компонент заказов
 
-Логический компонент для операций:
+Компонент обрабатывает операции:
 
 - `create_order`
 - `select_executor`
 - `auto_search_executor`
 - сценариев работы с `price_offer` и `confirm_price`
 
-Реализация пока остаётся in-process и использует существующие Go-пакеты из
-`internal/`.
+Код обработки входящих bus-сообщений находится в `component.go`.
+HTTP-часть и инфраструктурные зависимости пока остаются в `internal/`.
