+++ src/registry_component/README.md
@@ -1,9 +1,10 @@
 ## Компонент реестров
 
-Логический компонент для операций:
+Компонент обрабатывает операции:
 
 - `register_operator`
 - `register_customer`
 
-Реализация пока остаётся in-process и вызывается через gateway, а маршрутизация
-определена в `internal/gateway/gateway.go`.
+Код обработки входящих bus-сообщений находится в `component.go`.
+Вызов компонента идёт через gateway, а общий слой совместимости остаётся в
+`internal/handler`.
